fix: mostra i nomi nelle righe Temperatura configurate
