+++ custom_components/dashboardmodern/frontend/legacy/build-info.js
@@ -7,4 +7,4 @@ import "../src/sections/beta12-real-device-polish-section.js";
 import "../src/sections/room-color-lock-section.js";
 import "../src/sections/beta14-real-device-hotfix-section.js";
 import "../src/sections/beta16-real-device-layout-section.js";
-export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.0.0-beta.23","dashboardVersion":"1.0.0-beta.23","moduleVersion":14,"schemaVersion":4,"date":"2026-08-12T18:17:51+00:00","commit":"6e7303980cf17c2e2931e15878eab3c4045d2c0b","assetHash":"4c9e657b79fadb14"});
+export const BUILD_INFO = Object.freeze({"generated":true,"integrationVersion":"1.0.0-beta.24","dashboardVersion":"1.0.0-beta.24","moduleVersion":14,"schemaVersion":4,"date":"2026-08-12T18:17:51+00:00","commit":"6e7303980cf17c2e2931e15878eab3c4045d2c0b","assetHash":"4c9e657b79fadb14"});
